[zircon] Add support for booting A/B/R partitions based on sysconfig partition
Adds a new command zircon_bootconfig, which takes 4 arguments:

1. name of sysconfig partition
2. name of zircon-a partition
2. name of zircon-b partition
2. name of zircon-r partition

This command first attempts to read from the boot-oneshot section of the
sysconfig partition. If that section is empty, it then reads from the
boot-default section. Then based on the value of the "boot" key,
it sets the shell variable "boot_part" to the appropriate zircon-* partition name.
It then erases the boot-oneshot section if it was valid.

Change-Id: I104cbff39b0547e052209a77bbb20755b1ae47e7