[usb] version 0.09: Add hack to work around bad state induced by lsusb
lsusb requests device qualifier and debug descriptors, which the USB stack
in u-boot does not implement. But the dwc2 driver is incapable of stalling
endpoint zero, so USB gets in a bad state after that.
Here we work around the problem by returning dummy values for the
device qualifier and debug descriptors.