[vim2] Fix rebooting issue while booting.
There's an issue with some of the vim2 devices, in which if the device
is reset and it takes a long time between setting up the network chip
and initialising the Wake on Lan (WOL) feature, it would be turned off
and on. This issue only affects devices running MCU v5 and it doesn't
happen during cold boots.

This commit disables WOL related resets in the MCU during that time
windows. `kbi resetflag 0` is called right after starting the NIC.
The initialization code later sets the resetflag back to 1, but right
before that, we run `kbi forcereset wol w 0` to disable wol resets.
Finally, if wol was enabled, during `wol_init` we set the forcereset
flag to 1 again.

The workaround is kind of hacky because `kbi forcereset` cannot be
called before the nic is started (it writes something to the nic
registers), so I just used `kbi resetflag` in its place, but then
later in the code the nic initialization code sets the resetflag to
1... At that point `kbi forcereset` is available to use so we just
disable it, and then enable it again if WOL is also enabled.

Without this workaround, to trigger the bug it is enough to add a delay
(with `sleep 1` or `udelay(1000000)` anywhere in between the line that
connects the phyreset pin:

```c
setbits_le32(PREG_PAD_GPIO3_O, 1 << 14);
```

And the code that calls `kbi init`.

TEST=I have tested this in my vim2 device, and tried resetting the
device via WOL and it works.

ZX-2635 #Comment: Workaround disabling WOL resets.

Change-Id: Ibc3fd9817322e66244297e69f2fb7b02629ddffa